Refactor folders to separate concerns
