+++ client/journalapp/app/Entry_Flow/Entry.js
