+++ client/journalapp/app/Entry_Flow/EntryList.js
