+++ client/journalapp/app/Entry_Flow/MessageScene.js
