+++ client/journalapp/app/Friend_Flow/Friend.js
