+++ client/journalapp/app/Friend_Flow/Request.js
