+++ client/journalapp/app/Friend_Flow/RequestList.js
