+++ client/journalapp/app/Friend_Flow/SearchFriends.js
