+++ client/journalapp/app/Friend_Flow/SearchResultRow.js
