+++ client/journalapp/app/Friend_Flow/SearchResultsList.js
